fix(#351): Ensure CPUs is an integer value.

+++ gtdbtk/external/tigrfam_search.py
@@ -159,7 +159,7 @@ class TigrfamSearch(object):
         """
         if len(gene_files) == 0:
             raise GTDBTkExit('There are no genomes to process.')
-        self.cpus_per_genome = max(1, self.threads / len(gene_files))
+        self.cpus_per_genome = max(1, int(self.threads / len(gene_files)))
 
         # populate worker queue with data to process
         workerQueue = mp.Queue()
